returns double with 4 digits after decimal point

+++ src/main/kotlin/com/epam/cdp/caclulator/Operation.kt
@@ -1,3 +1,3 @@
 package com.epam.cdp.caclulator
 
-data class Operation(val operator: String, val operand1: Int, val operand2: Int)
+data class Operation(val operator: String, val operand1: Double, val operand2: Double)
